test_generators/ssz/requirements.txt
ruamel.yaml==0.15.87
ssz==0.1.0a2
